+++ src/scripts/index.js
@@ -1,6 +1,8 @@
-import 'bootstrap/dist/js/bootstrap.bundle.js';
+import bootstrap from 'bootstrap/dist/js/bootstrap.bundle.js';
+window.bootstrap = bootstrap;
 
 import './components/mobile-controls.js';
 import './components/return-top.js';
 import './components/external-link.js';
 import './components/header.js';
+import './components/navigation.js';