+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.19.19
+- Updated npm packages
+- bootstrap/dist/js/bootstrap.bundle.js added to window.bootstrap global variable
+- Added script for navigation
+- Update megadropdown navigation styles
+
 v1.19.18 
 - Further refined MutationObserver/ResizeObserver recalculates when size changes occur above the header as well now